competitionsService-Tiny/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
tinydb==4.8.2
python-dotenv==1.0.1
